PJT2_DB_cx_Oracle/requirements.txt
beautifulsoup4
requests
cx_Oracle
pillow
